chatserver 10:03 10/3/24

+++ Chat_Server/src/main/java/com/chat/websocket/repository/GroupMemberRepository.java
@@ -14,5 +14,5 @@ public interface GroupMemberRepository extends JpaRepository<GroupMember,Long> {
     Optional<GroupMember> findGroupMemberByConversationId(@Param("conversationId") long conversationId);
 
     @Query("select gm from GroupMember gm where gm.contact.email = :email and gm.conversation.id = :conversationId" )
-    Optional<GroupMember> findByEmailAndConversationId(@Param("email") String email,@Param("conversationId") int conversationId);
+    Optional<GroupMember> findByEmailAndConversationId(@Param("email") String email,@Param("conversationId") long conversationId);
 }
